Move Ranks to its own module.
